[Change] Multiple QOL Changes (See Details)
- `ERROR_BYTE_OVERFLOW` changed to `ERROR_BYTE_INSUFFICIENT` to represent underflow cases when using `PaintSomething`
- `EVENT_SKIPPED` and `EVENT_PROCESSING` text has been rewritten to be presented in a cleaner format.
- Removed `DispatchModernUpdate` and `DispatchLegacyUpdate`. The converged `DispatchUpdate` now uses the new vararg method to parse arguments with pre-filled arguments up to 10 given for 1.12.1 and below users.

Notes: I am planning in a future update to allow customization/adding/removing of `ignore_event_conditions`, with it also being added as a optional parameter within `ModifyTriggers`. This is mainly so that those using custom placeholders are not limited by the static event-conditions set in place as well as can define their own for any events needed for their placeholders. Current ETA is before `v1.1.5`